+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### vNext
 
-
+- [feature] Add CERE token transfer SDK method and in-wallet UI
 
 ### v1.14.0